Refactored the Indigo extension file
