+++ indigo-sandboxes/perf/src/com/example/perf/Runtime.scala
@@ -10,7 +10,7 @@ object Runtime extends BasicGameRuntime:
   def game: Game[?, ?, ?] =
     PerfGame()
 
-  def settings: Indigo.Settings =
-    Indigo.Settings.default
+  def settings: Settings =
+    Settings.default
       .withFrameRatePolicy(FrameRatePolicy.Unlimited)
       .allowContextMenu
