+++ indigo-sandboxes/sandbox/src/main/scala/com/example/sandbox/Runtime.scala
@@ -10,6 +10,6 @@ object Runtime extends BasicGameRuntime:
   def game: Game[?, ?, ?] =
     SandboxGame()
 
-  def settings: Indigo.Settings =
-    Indigo.Settings.default
+  def settings: Settings =
+    Settings.default
       .withFrameRatePolicy(FrameRatePolicy.Skip(FPS.`60`))
