+++ indigo/src-js/indigo/internal/IndigoActions.scala
@@ -1,0 +1,38 @@
+package indigo.internal
+
+import indigo.*
+import indigo.internal.models.LaunchStatus
+import indigo.internal.models.Msg
+import indigo.platform.IndigoCoreServices
+import indigo.render.facades.WebGL2RenderingContext
+import org.scalajs.dom.html
+import tyrian.*
+import tyrian.extensions.ExtensionId
+
+object IndigoActions:
+
+  def launch(
+      extensionId: ExtensionId,
+      game: Game[?, ?, ?],
+      maybeCanvas: Option[html.Canvas],
+      flags: Map[String, String],
+      settings: Settings,
+      services: IndigoCoreServices
+  ): Action =
+    Action.run {
+      maybeCanvas match
+        case Some(canvas) =>
+          val context: WebGL2RenderingContext =
+            CanvasAndContext.setupContext(
+              canvas,
+              settings.premultipliedAlpha,
+              settings.transparentBackground,
+              settings.antiAliasing
+            )
+
+          game.launch(canvas.width, canvas.height, context, flags, services)
+          Msg.Launch(LaunchStatus.Started(extensionId))
+
+        case _ =>
+          Msg.Launch(LaunchStatus.Retry(extensionId))
+    }
