+++ indigo/src-js/indigo/internal/models/FullScreenRequest.scala
@@ -1,0 +1,4 @@
+package indigo.internal.models
+
+enum FullScreenRequest derives CanEqual:
+  case Enter, Exit, Toggle
