+++ indigo/src-js/indigo/internal/models/TickUpdateResult.scala
@@ -1,0 +1,7 @@
+package indigo.internal.models
+
+import indigo.Seconds
+
+enum TickUpdateResult derives CanEqual:
+  case Wait
+  case RunNow(timeDelta: Seconds, updatedAt: Seconds)
