+++ indigo/test/src-js/indigo/internal/UtilsTests.scala
@@ -1,6 +1,11 @@
-package indigo
+package indigo.internal
 
-class IndigoTests extends munit.FunSuite:
+import indigo.FPS
+import indigo.FrameRatePolicy
+import indigo.Seconds
+import indigo.internal.models.TickUpdateResult
+
+class UtilsTests extends munit.FunSuite:
 
   val lastUpdated: Seconds             = Seconds(10)
   val frameRatePolicy: FrameRatePolicy = FrameRatePolicy.Skip(FPS(60))
@@ -9,10 +14,10 @@ class IndigoTests extends munit.FunSuite:
     val runningTime: Seconds = Seconds(10.032) // Two frames worth of time
 
     val actual =
-      Indigo.processFrameTick(lastUpdated, runningTime, frameRatePolicy)
+      Utils.processFrameTick(lastUpdated, runningTime, frameRatePolicy)
 
-    val expected: Indigo.TickUpdateResult =
-      Indigo.TickUpdateResult.RunNow(Seconds(0.032), runningTime)
+    val expected: TickUpdateResult =
+      TickUpdateResult.RunNow(Seconds(0.032), runningTime)
 
     assert(resultCloseEnough(clue(actual.unsafeGet), clue(expected)))
   }
@@ -21,21 +26,21 @@ class IndigoTests extends munit.FunSuite:
     val runningTime: Seconds = Seconds(10.001)
 
     val actual =
-      Indigo.processFrameTick(lastUpdated, runningTime, frameRatePolicy)
+      Utils.processFrameTick(lastUpdated, runningTime, frameRatePolicy)
 
-    val expected: Indigo.TickUpdateResult =
-      Indigo.TickUpdateResult.Wait
+    val expected: TickUpdateResult =
+      TickUpdateResult.Wait
 
     assert(resultCloseEnough(clue(actual.unsafeGet), clue(expected)))
   }
 
   // JavaScript floating point precision comparison helper
-  def resultCloseEnough(a: Indigo.TickUpdateResult, b: Indigo.TickUpdateResult): Boolean =
+  def resultCloseEnough(a: TickUpdateResult, b: TickUpdateResult): Boolean =
     (a, b) match
-      case (Indigo.TickUpdateResult.Wait, Indigo.TickUpdateResult.Wait) =>
+      case (TickUpdateResult.Wait, TickUpdateResult.Wait) =>
         true
 
-      case (Indigo.TickUpdateResult.RunNow(tdA, rtA), Indigo.TickUpdateResult.RunNow(tdB, rtB)) =>
+      case (TickUpdateResult.RunNow(tdA, rtA), TickUpdateResult.RunNow(tdB, rtB)) =>
         closeEnough(tdA.toDouble, tdB.toDouble) && closeEnough(rtA.toDouble, rtB.toDouble)
 
       case _ =>
